+++ src/main/java/io/github/computerdaddyguy/jfiletreeprettyprinter/scanner/ScanningOptions.java
@@ -5,7 +5,6 @@ import java.util.Comparator;
 import java.util.function.Predicate;
 import java.util.function.ToIntFunction;
 import org.jspecify.annotations.NullMarked;
-import org.jspecify.annotations.Nullable;
 
 @NullMarked
 public interface ScanningOptions {
@@ -16,7 +15,6 @@ public interface ScanningOptions {
 
 	Comparator<Path> pathComparator();
 
-	@Nullable
 	Predicate<Path> pathFilter();
 
 }